requirements.txt
# Python 3.12.3
blinker==1.8.1
click==8.1.7
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
python-dotenv==1.0.1
SQLAlchemy==2.0.29
Werkzeug==3.0.2
requests==2.32.2